refactor: share network policy conversion helpers
